Moving things around in rpc package
